+++ style_generator/lib/src/annotations/copy_with_key_internal.dart
@@ -1,23 +1,37 @@
 /// @docImport "package:style_generator_annotation/copy_with_generator_annotation.dart";
 library;
 
+import "package:meta/meta.dart";
+
 /// The internal representation of [CopyWithKey]
-class CopyWithKeyInternal<T> {
+class CopyWithKeyInternal {
   static const String srcAnnotationName = "CopyWithKey";
   static const String inCopyWithName = "inCopyWith";
+  static const String unmodifiableName = "unmodifiable";
   static const String fieldName = "field";
 
   final bool inCopyWith;
+  final bool unmodifiable;
   final String? field;
 
   const CopyWithKeyInternal({
     required this.inCopyWith,
+    required this.unmodifiable,
     required this.field,
   });
 
+  /// These defaults must be the same as the one set by [CopyWithKey]
+  @internal
+  static const CopyWithKeyInternal defaults = CopyWithKeyInternal(
+    inCopyWith: true,
+    unmodifiable: false,
+    field: null,
+  );
+
   factory CopyWithKeyInternal.fromJson(Map<String, Object?> json) {
     return CopyWithKeyInternal(
-      inCopyWith: json[inCopyWithName] as bool? ?? true,
+      inCopyWith: json[inCopyWithName]! as bool,
+      unmodifiable: json[unmodifiableName]! as bool,
       field: json[fieldName] as String?,
     );
   }
@@ -25,6 +39,7 @@ class CopyWithKeyInternal<T> {
   Map<String, Object?> toJson() {
     return {
       inCopyWithName: inCopyWith,
+      unmodifiableName: unmodifiable,
       fieldName: field,
     };
   }